+++ ixmp/tests/test_cli.py
@@ -242,3 +242,44 @@ def test_report(ixmp_cli):
     # 'report' without specifying a platform/scenario is a UsageError
     result = ixmp_cli.invoke(['report', 'key'])
     assert result.exit_code == UsageError.exit_code
+
+
+def test_solve(ixmp_cli, test_mp):
+    populate_test_platform(test_mp)
+    cmd = [
+        '--platform', test_mp.name,
+        '--model', models['dantzig']['model'],
+        '--scenario', models['dantzig']['scenario'],
+        'solve', '--remove-solution'
+    ]
+    result = ixmp_cli.invoke(cmd)
+    assert result.exit_code == 0, result.output
+
+    # test failing solving without solution removal
+    cmd = [
+        '--platform', test_mp.name,
+        '--model', models['dantzig']['model'],
+        '--scenario', models['dantzig']['scenario'],
+        'solve'
+    ]
+    result = ixmp_cli.invoke(cmd)
+    assert result.exit_code == 1, result.output
+
+    # missing scenario
+    cmd = [
+        '--platform', test_mp.name,
+        '--model', 'non-existing',
+        '--scenario', models['dantzig']['scenario'],
+        'solve'
+    ]
+    result = ixmp_cli.invoke(cmd)
+    assert result.exit_code == 1, result.output
+    assert "Error: model='non-existing'" in result.output
+
+    # no platform/scenario provided
+    cmd = [
+        'solve'
+    ]
+    result = ixmp_cli.invoke(cmd)
+    assert result.exit_code != 0, result.output
+    assert "Error: give --url before command solve" in result.output